fix(package): Add a CommonJS ES5 build in ./lib
This commit adds a build step for CommonJS files to be consumed via Browserify and Webpack.
It changes `main` to point to `lib/index.js` because this is the CommonJS entry point.
This commit fixes "Cannot find module 'stackframe'" when compiling with Browserify (https://github.com/KeywordBrain/redbox-react/issues/17#issuecomment-140860875).